+++ pkg/deployment/deployment.go
@@ -136,6 +136,14 @@ type Deployment struct {
 	haveServiceMonitorCRD     bool
 
 	memberState memberState.StateInspector
+
+	metrics struct {
+		agency struct {
+			errors  uint64
+			fetches uint64
+			index   uint64
+		}
+	}
 }
 
 func (d *Deployment) WithArangoMember(cache inspectorInterface.Inspector, timeout time.Duration, name string) reconciler.ArangoMemberModContext {