+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - (Feature) Change DBServer Cleanup Logic
 - (Feature) Set Logger format
 - (Bugfix) Ensure Wait actions to be present after AddMember
+- (Documentation) Refactor metrics (Part 1)
 
 ## [1.2.13](https://github.com/arangodb/kube-arangodb/tree/1.2.13) (2022-06-07)
 - (Bugfix) Fix arangosync members state inspection